feat: Add InfoWindow when user clicks on the Marker or list
Now when the user clicks on a marker or entry of places list,
we fetch some snippet of the place from Wikipedia, and
set it as InfoWindow content. If no data is found,
an error message is set.
Added new methods to the model, to control InfoWindows
visibility.
New map center location set, because we had better search
results, and thus give us a better scenario for testing.